fix: update type, split components + simplify sort
